Extract sizing into its own file
