+++ app/src/ui/diff/image-diffs/modified-image-diff.tsx
@@ -8,17 +8,13 @@ import { DifferenceBlend } from './difference-blend'
 import { OnionSkin } from './onion-skin'
 import { Swipe } from './swipe'
 import { assertNever } from '../../../lib/fatal-error'
+import { IImageSize, getMaxFitSize } from './sizing'
 
 interface IModifiedImageDiffProps {
   readonly previous: Image
   readonly current: Image
   readonly diffType: ImageDiffType
   readonly onChangeDiffType: (type: ImageDiffType) => void
-}
-
-export interface IImageSize {
-  readonly width: number
-  readonly height: number
 }
 
 export interface ICommonImageDiffProperties {
@@ -37,51 +33,6 @@ interface IModifiedImageDiffState {
 
   /** The size of the current image. */
   readonly currentImageSize: IImageSize | null
-}
-
-/**
- * Get the size which fits in the container without scaling and maintaining
- * aspect ratio.
- */
-function getAspectFitSize(
-  imageSize: IImageSize,
-  containerSize: IImageSize
-): IImageSize {
-  const heightRatio =
-    containerSize.height < imageSize.height
-      ? imageSize.height / containerSize.height
-      : 1
-  const widthRatio =
-    containerSize.width < imageSize.width
-      ? imageSize.width / containerSize.width
-      : 1
-
-  let ratio = Math.max(1, widthRatio)
-  if (widthRatio < heightRatio) {
-    ratio = Math.max(1, heightRatio)
-  }
-
-  return {
-    width: imageSize.width / ratio,
-    height: imageSize.height / ratio,
-  }
-}
-
-/**
- * Get the size which will fit the bigger of the two images while maintaining
- * aspect ratio.
- */
-function getMaxFitSize(
-  previousImageSize: IImageSize,
-  currentImageSize: IImageSize,
-  containerSize: IImageSize
-): IImageSize {
-  const previousSize = getAspectFitSize(previousImageSize, containerSize)
-  const currentSize = getAspectFitSize(currentImageSize, containerSize)
-
-  const width = Math.max(previousSize.width, currentSize.width)
-  const height = Math.max(previousSize.height, currentSize.height)
-  return { width, height }
 }
 
 /** A component which renders the changes to an image in the repository */
